Feature ETP-87: Move login form to component library.
